Add lots of missing std's.
git-svn-id: https://svn.dealii.org/trunk@8487 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/deal.II/dof_renumbering.cc
@@ -86,7 +86,7 @@ check_renumbering(MGDoFHandler<dim>& mgdof, bool discontinuous)
 
   Point<dim> direction;
   for (unsigned int i=0;i<dim;++i)
-    direction(i) = pow(10.,i);
+    direction(i) = std::pow(10.,static_cast<double>(i));
   
 				   // Check global ordering
   print_dofs (dof);
